requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
numpy
pillow
pandas
python-multipart